requirements.txt
yt-dlp>=2024.8.0
faster-whisper>=1.0.3
ffmpeg-python>=0.2.0
ollama>=0.3.3
groq>=0.11.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
loguru>=0.7.2
python-dotenv>=1.0.1
pydantic>=2.7.0
jinja2>=3.1.4
